frontend: split repo feature into modules
