--- Frontend/src/scripts/features/repo.ts
@@ -1,1368 +1,0 @@
-// src/scripts/features/repo.ts
-import { qs, qsa, escapeHtml } from '../lib/dom';
-import { buildCtxMenu, CtxItem } from '../lib/menu';
-import { TAURI } from '../lib/tauri';
-import { notify } from '../lib/notify';
-import { state, prefs, statusLabel, statusClass } from '../state/state';
-import { refreshRepoActions } from '../ui/layout';
-import { openStashConfirm } from './stashConfirm';
-
-const filterInput   = qs<HTMLInputElement>('#filter');
-const selectAllBox  = qs<HTMLInputElement>('#select-all');
-const listEl        = qs<HTMLElement>('#file-list');
-const countEl       = qs<HTMLElement>('#changes-count');
-const leftFootEl    = qs<HTMLElement>('#left-foot');
-const undoLeftBtn   = leftFootEl?.querySelector<HTMLButtonElement>('#undo-left-btn') ?? null;
-let stashFootEl: HTMLElement | null = null;
-let stashFootBound = false;
-
-const diffHeadPath  = qs<HTMLElement>('#diff-path');
-const diffEl        = qs<HTMLElement>('#diff');
-let lastClickedIndex: number = -1;
-let isDragSelecting = false;
-let dragTargetState = true; // true=select, false=deselect
-let dragVisited = new Set<string>();
-let dragMoved = false;
-let suppressNextClick = false;
-let dragMode: 'diff' | 'commit' | null = null;
-let dragStartIndex: number = -1;
-let dragCurrentIndex: number = -1;
-let dragPreDiff = new Set<string>();
-let dragPrePicked = new Set<string>();
-
-// Global guards to suppress native text selection/drag while we paint-select
-document.addEventListener('selectstart', (e) => { if (isDragSelecting) e.preventDefault(); }, true);
-document.addEventListener('dragstart',   (e) => { if (isDragSelecting) e.preventDefault(); }, true);
-
-export function bindRepoHotkeys(commitBtn: HTMLButtonElement | null, openSheet: (w: 'clone'|'add'|'switch') => void) {
-    if (!filterInput) return;
-    window.addEventListener('keydown', (e) => {
-        const key = e.key.toLowerCase();
-        if (e.ctrlKey && key === 'f') { e.preventDefault(); filterInput.focus(); }
-        if (e.ctrlKey && key === 'r') { e.preventDefault(); openSheet('switch'); }
-        if (e.ctrlKey && e.key === 'Enter') { e.preventDefault(); commitBtn?.click(); }
-        if (e.key === 'Escape') {
-            const about = document.getElementById('about-modal');
-            if (about?.classList.contains('show')) about.classList.remove('show');
-            // do not force-open the command sheet on Esc; let modal system handle closes
-        }
-    });
-}
-
-export function bindFilter() {
-    filterInput?.addEventListener('input', () => renderList());
-    selectAllBox?.addEventListener('change', () => {
-        if (prefs.tab !== 'changes') return;
-        state.defaultSelectAll = false;
-        const files = getVisibleFiles();
-        toggleSelectAll(Boolean(selectAllBox?.checked), files);
-        renderList();
-    });
-}
-
-export function renderList() {
-    if (!listEl || !countEl || !filterInput || !diffHeadPath || !diffEl) return;
-
-    listEl.innerHTML = '';
-    const isHistory = prefs.tab === 'history';
-    const isStash = prefs.tab === 'stash';
-    // Toggle list styling for history/stash vs changes
-    if (isHistory || isStash) listEl.classList.add('commit-list');
-    else listEl.classList.remove('commit-list');
-    const q = filterInput.value.trim().toLowerCase();
-    updateCommitButton();
-
-    if (isStash) showStashFooter();
-    else hideStashFooter();
-    refreshRepoActions();
-
-    if (isHistory) {
-        const commits = (state.commits || []).filter(c =>
-            !q || c.msg?.toLowerCase().includes(q) || c.id?.includes(q)
-        );
-
-        countEl.textContent = `${commits.length} commit${commits.length === 1 ? '' : 's'}`;
-
-        if (!commits.length) {
-            listEl.innerHTML = `<li class="row" aria-disabled="true"><div class="file">No commits loaded.</div></li>`;
-            diffHeadPath.textContent = 'Commit details';
-            diffEl.innerHTML = '';
-            return;
-        }
-
-        // Show an informational row for commits present on remote but not pulled locally
-        const ahead = Number((state as any).ahead || 0);
-        const behind = Number((state as any).behind || 0);
-        if (behind > 0) {
-            const info = document.createElement('li');
-            info.className = 'row notice';
-            info.innerHTML = `<div class="file" title="Commits exist on the remote that are not pulled locally">↓ ${behind} incoming commit${behind === 1 ? '' : 's'} on remote</div>`;
-            listEl.appendChild(info);
-        }
-
-        const aheadIds: Set<string> = (state as any).aheadIds || new Set<string>();
-        commits.forEach((c, i) => {
-            const li = document.createElement('li');
-            li.className = 'row commit';
-            const short = (c.id || '').slice(0, 7);
-            const whenRaw = String(c.meta || '').split('•')[0].trim();
-            const rel = formatTimeAgo(whenRaw);
-            const exact = (c.meta || '').trim();
-            const isAhead = !!(c?.id && (aheadIds.size > 0 ? aheadIds.has(c.id) : (i < ahead)));
-            const statusTag = isAhead ? `<span class=\"tag up\" title=\"Not on remote yet\">↑ outgoing</span>` : '';
-            li.innerHTML = `
-        <span class="badge hash" title="${escapeHtml(c.id || '')}">${escapeHtml(short)}</span>
-        <div class="file" title="${escapeHtml(c.msg || '')}">${escapeHtml(c.msg || '(no message)')}</div>
-        ${statusTag}
-        <span class="badge time" title="${escapeHtml(exact)}">${escapeHtml(rel)}</span>`;
-            li.addEventListener('click', () => selectHistory(c, i));
-            // Right-click commit → context menu (includes Undo when unpushed)
-            li.addEventListener('contextmenu', (ev) => {
-                ev.preventDefault();
-                const x = (ev as MouseEvent).clientX, y = (ev as MouseEvent).clientY;
-                const items: CtxItem[] = [];
-                // Always offer copy hash for discoverability so menu shows
-                items.push({ label: 'Copy hash', action: async () => {
-                    try {
-                        await navigator.clipboard.writeText(c.id || '');
-                        notify('Hash copied');
-                    } catch { /* ignore */ }
-                }});
-                if (isAhead) {
-                    items.push({ label: '---' });
-                    items.push({ label: 'Undo to this commit', action: async () => {
-                        if (!TAURI.has) return;
-                        try {
-                            await TAURI.invoke('git_undo_to_commit', { id: c.id });
-                            await Promise.allSettled([hydrateStatus(), hydrateCommits()]);
-                        } catch { notify('Undo failed'); }
-                    }});
-                }
-                buildCtxMenu(items, x, y);
-            });
-            listEl.appendChild(li);
-        });
-        selectHistory(commits[0], 0);
-        return;
-    }
-
-    // Stash tab rendering
-    if (isStash) {
-        const stash = ((state as any).stash || []) as any[];
-        const items = stash.filter((s: any) => !q || (s.msg || '').toLowerCase().includes(q) || (s.selector || '').includes(q));
-        countEl.textContent = `${items.length} stash${items.length === 1 ? '' : 'es'}`;
-
-        const enableActionButtons = (enabled: boolean) => {
-            const a = qs<HTMLButtonElement>('#stash-apply-btn'); if (a) a.disabled = !enabled;
-            const p = qs<HTMLButtonElement>('#stash-pop-btn'); if (p) p.disabled = !enabled;
-            const d = qs<HTMLButtonElement>('#stash-drop-btn'); if (d) d.disabled = !enabled;
-        };
-
-        if (!items.length) {
-            listEl.innerHTML = `<li class="row" aria-disabled="true"><div class="file">No stashes.</div></li>`;
-            diffHeadPath.textContent = 'Stash details';
-            diffEl.innerHTML = '';
-            enableActionButtons(false);
-            return;
-        }
-
-        items.forEach((s: any, i: number) => {
-            const li = document.createElement('li');
-            li.className = 'row commit';
-            const sel = s.selector || '';
-            const exact = (s.meta || '').trim();
-            li.dataset.selector = sel;
-            li.innerHTML = `
-        <div class="file" title="${escapeHtml(s.msg || '')}">${escapeHtml(s.msg || '(no message)')}</div>
-        <span class="badge time" title="${escapeHtml(exact)}">${escapeHtml(exact)}</span>`;
-            li.addEventListener('click', () => selectStash(s, i));
-            li.addEventListener('contextmenu', (ev) => {
-                ev.preventDefault();
-                state.currentStash = sel;
-                enableActionButtons(true);
-                const mev = ev as MouseEvent;
-                const x = mev.clientX, y = mev.clientY;
-                const target = sel;
-                const items: CtxItem[] = [];
-                items.push({ label: 'Apply stash', action: async () => {
-                    try {
-                        if (!TAURI.has) return;
-                        await TAURI.invoke('git_stash_apply', { selector: target });
-                        notify('Applied stash');
-                        await Promise.allSettled([hydrateStatus(), hydrateStash()]);
-                        renderList();
-                    } catch { notify('Failed to apply stash'); }
-                }});
-                items.push({ label: 'Delete stash', action: async () => {
-                    const ok = window.confirm(`Delete ${target}? This cannot be undone.`);
-                    if (!ok) return;
-                    try {
-                        if (!TAURI.has) return;
-                        await TAURI.invoke('git_stash_drop', { selector: target });
-                        notify('Deleted stash');
-                        if (state.currentStash === target) state.currentStash = '';
-                        await Promise.allSettled([hydrateStash()]);
-                        renderList();
-                    } catch { notify('Failed to delete stash'); }
-                }});
-                buildCtxMenu(items, x, y);
-            });
-            listEl.appendChild(li);
-        });
-        selectStash(items[0], 0);
-        return;
-    }
-
-    const files = (state.files || []).filter(f =>
-        !q || (f.path || '').toLowerCase().includes(q)
-    );
-    countEl.textContent = `${files.length} file${files.length === 1 ? '' : 's'}`;
-    updateSelectAllState(files);
-
-    if (!files.length) {
-        listEl.innerHTML = `<li class="row" aria-disabled="true"><div class="file">No changes. Clone or add a repository to get started.</div></li>`;
-        diffHeadPath.textContent = 'Select a file to view changes';
-        diffEl.innerHTML = '';
-        updateSelectAllState([]);
-        updateCommitButton();
-        return;
-    }
-
-    files.forEach((f, i) => {
-        const li = document.createElement('li');
-        li.className = 'row';
-        li.setAttribute('role', 'option');
-        li.setAttribute('data-path', f.path || '');
-        const picked = state.selectedFiles.has(f.path);
-        const diffsel = state.diffSelectedFiles.has(f.path);
-        li.classList.toggle('picked', picked);
-        li.classList.toggle('diffsel', diffsel);
-        li.innerHTML = `
-      <input type="checkbox" class="pick" aria-label="Select file" ${picked ? 'checked' : ''} />
-      <span class="status ${statusClass(f.status)}">${escapeHtml(f.status || '')}</span>
-      <div class="file" title="${escapeHtml(f.path || '')}">${escapeHtml(f.path || '')}</div>
-      <span class="pick-mark" aria-hidden="true">✓</span>`;
-        li.addEventListener('click', (e) => onFileClick(e as MouseEvent, f, i, files));
-        li.addEventListener('mousedown', (e) => onFileMouseDown(e as MouseEvent, f, i, files, li));
-        li.addEventListener('mouseenter', () => {
-            if (isDragSelecting) {
-                dragCurrentIndex = i;
-                updateDragRange(files);
-            }
-        });
-        li.addEventListener('contextmenu', (ev) => onFileContextMenu(ev, f));
-        // prevent row click when toggling checkbox
-        const cb = li.querySelector<HTMLInputElement>('input.pick');
-        if (cb) cb.dataset.path = f.path || '';
-        cb?.addEventListener('click', (ev) => {
-            ev.stopPropagation();
-            toggleFilePick(f.path, !!cb?.checked);
-            updateSelectAllState(files);
-            li.classList.toggle('picked', !!cb?.checked && !(cb as any).indeterminate);
-        });
-        listEl.appendChild(li);
-    });
-
-    // Preserve current viewed file if present; otherwise show combined diff or first file
-    const curIdx = state.currentFile ? files.findIndex(x => x.path === state.currentFile) : -1;
-    if (state.diffSelectedFiles && state.diffSelectedFiles.size > 1) {
-        renderCombinedDiff(Array.from(state.diffSelectedFiles));
-    } else if (curIdx >= 0) {
-        selectFile(files[curIdx], curIdx);
-    } else {
-        selectFile(files[0], 0);
-    }
-    updateCommitButton();
-}
-
-function highlightRow(index: number) {
-    const sel = (prefs.tab === 'history') ? '.row.commit' : '.row';
-    const rows = qsa<HTMLElement>(sel, listEl || (undefined as any));
-    rows.forEach((el, i) => el.classList.toggle('active', i === index));
-}
-
-async function selectFile(file: { path: string }, index: number) {
-    if (!diffHeadPath || !diffEl) return;
-    highlightRow(index);
-    diffHeadPath.textContent = file.path || '(unknown file)';
-    diffEl.innerHTML = '<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Loading…</div></div></div>';
-
-    try {
-        let lines: string[] = [];
-        if (TAURI.has && file.path) {
-            lines = await TAURI.invoke<string[]>('git_diff_file', { path: file.path });
-        }
-        state.currentFile = file.path;
-        state.currentDiff = lines || [];
-        diffEl.innerHTML = renderHunksWithSelection(state.currentDiff);
-        bindHunkToggles(diffEl);
-        // Right-click on hunk → context menu to discard this hunk
-        const onCtx = (ev: Event) => {
-            const mev = ev as MouseEvent;
-            if (mev.type !== 'contextmenu') return;
-            const hk = (mev.target as HTMLElement).closest('.hunk') as HTMLElement | null;
-            if (!hk) return;
-            mev.preventDefault();
-            const idxAttr = hk.getAttribute('data-hunk-index');
-            const hi = idxAttr ? Number(idxAttr) : -1;
-            if (hi < 0) return;
-            const x = mev.clientX, y = mev.clientY;
-            const items: { label: string; action: () => void }[] = [];
-            items.push({ label: 'Discard hunk', action: async () => {
-                if (!TAURI.has) return;
-                const ok = window.confirm('Discard this hunk? This cannot be undone.');
-                if (!ok) return;
-                try {
-                    const patch = buildPatchForSelectedHunks(file.path, state.currentDiff, [hi]);
-                    if (patch) {
-                        await TAURI.invoke('git_discard_patch', { patch });
-                        await Promise.allSettled([hydrateStatus()]);
-                    }
-                } catch { notify('Discard failed'); }
-            }});
-            const selected = (state as any).selectedHunksByFile?.[file.path] as number[] | undefined;
-            if (Array.isArray(selected) && selected.length > 0) {
-                items.push({ label: 'Discard selected hunks (this file)', action: async () => {
-                    if (!TAURI.has) return;
-                    const ok = window.confirm(`Discard ${selected.length} selected hunk(s) in this file? This cannot be undone.`);
-                    if (!ok) return;
-                    try {
-                        const patch = buildPatchForSelectedHunks(file.path, state.currentDiff, selected);
-                        if (patch) {
-                            await TAURI.invoke('git_discard_patch', { patch });
-                            await Promise.allSettled([hydrateStatus()]);
-                        }
-                    } catch { notify('Discard failed'); }
-                }});
-            }
-            // Discard selected hunks across files
-            const hunksMap: Record<string, number[]> = (state as any).selectedHunksByFile || {};
-            const filesWithSel = Object.keys(hunksMap).filter(k => Array.isArray(hunksMap[k]) && hunksMap[k].length > 0);
-            if (filesWithSel.length > 0) {
-                items.push({ label: 'Discard selected hunks (all files)', action: async () => {
-                    if (!TAURI.has) return;
-                    const ok = window.confirm(`Discard selected hunks across ${filesWithSel.length} file(s)? This cannot be undone.`);
-                    if (!ok) return;
-                    try {
-                        let patch = '';
-                        for (const p of filesWithSel) {
-                            let lines: string[] = [];
-                            try { lines = await TAURI.invoke<string[]>('git_diff_file', { path: p }); } catch {}
-                            if (!Array.isArray(lines) || lines.length === 0) continue;
-                            patch += buildPatchForSelectedHunks(p, lines, hunksMap[p]) + '\n';
-                        }
-                        if (patch.trim()) {
-                            await TAURI.invoke('git_discard_patch', { patch });
-                            await Promise.allSettled([hydrateStatus()]);
-                        }
-                    } catch { notify('Discard failed'); }
-                }});
-            }
-            buildCtxMenu(items, x, y);
-        };
-        diffEl.addEventListener('contextmenu', onCtx, { once: true });
-        const cached = (state as any).selectedHunksByFile?.[file.path] as number[] | undefined;
-        if (Array.isArray(cached)) {
-            state.selectedHunks = cached.slice();
-            updateHunkCheckboxes();
-        } else if (state.selectedFiles.has(file.path) || state.defaultSelectAll) {
-            // Default-select all hunks and all changed lines for the current file
-            state.selectedHunks = allHunkIndices(state.currentDiff);
-            updateHunkCheckboxes();
-            // Seed per-line selections if not already present
-            const recExisting: Record<number, number[]> = (state as any).selectedLinesByFile[state.currentFile] || {};
-            const root = diffEl as HTMLElement;
-            const rec: Record<number, number[]> = { ...recExisting };
-            state.selectedHunks.forEach((h) => {
-                if (rec[h] && rec[h].length > 0) return;
-                const boxes = root.querySelectorAll<HTMLInputElement>(`input.pick-line[data-hunk="${h}"]`);
-                const picked: number[] = [];
-                boxes.forEach(b => {
-                    b.checked = true;
-                    picked.push(Number(b.dataset.line || -1));
-                });
-                if (picked.length > 0) rec[h] = Array.from(new Set(picked)).sort((a,b)=>a-b);
-            });
-            (state as any).selectedLinesByFile[state.currentFile] = rec;
-            updateHunkCheckboxes();
-        } else {
-            state.selectedHunks = [];
-        }
-        syncFileCheckboxWithHunks();
-        updateCommitButton();
-    } catch (e) {
-        console.error(e);
-        diffEl.innerHTML = `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Failed to load diff</div></div></div>`;
-    }
-}
-
-async function selectStash(item: { selector: string; msg?: string; meta?: string }, index: number) {
-    if (!diffHeadPath || !diffEl) return;
-    highlightRow(index);
-    state.currentStash = item.selector;
-    const title = (item.msg || '').trim();
-    diffHeadPath.textContent = title || item.selector;
-    diffEl.innerHTML = '<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Loading…</div></div></div>';
-    try {
-        let lines: string[] = [];
-        if (TAURI.has && item.selector) {
-            lines = await TAURI.invoke<string[]>('git_stash_show', { selector: item.selector });
-        }
-        const a = qs<HTMLButtonElement>('#stash-apply-btn'); if (a) a.disabled = false;
-        const p = qs<HTMLButtonElement>('#stash-pop-btn'); if (p) p.disabled = false;
-        const d = qs<HTMLButtonElement>('#stash-drop-btn'); if (d) d.disabled = false;
-        state.currentDiff = lines || [];
-        diffEl.innerHTML = renderHunksReadonly(state.currentDiff);
-    } catch (e) {
-        console.warn('git_stash_show failed', e);
-        diffEl.innerHTML = '<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Failed to load stash diff</div></div></div>';
-    }
-}
-
-async function selectHistory(commit: any, index: number) {
-    if (!diffHeadPath || !diffEl) return;
-    highlightRow(index);
-    const id = (commit.id || '').slice(0,7);
-    diffHeadPath.textContent = `Commit ${id || '(unknown)'}`;
-    diffEl.innerHTML = `
-    <div class="hunk">
-      <div class="hline"><div class="gutter">commit</div><div class="code">${escapeHtml(commit.id || '')}</div></div>
-      <div class="hline"><div class="gutter">Author</div><div class="code">${escapeHtml(commit.author || 'You <you@example.com>')}</div></div>
-      <div class="hline"><div class="gutter">Message</div><div class="code">${escapeHtml(commit.msg || '')}</div></div>
-    </div>
-    <div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Loading diff…</div></div></div>`;
-
-    // Load commit diff and render below the header
-    try {
-        let lines: string[] = [];
-        if (TAURI.has && commit.id) {
-            lines = await TAURI.invoke<string[]>('git_diff_commit', { id: commit.id });
-        }
-        const files = parseCommitDiffByFile(lines || []);
-        if (files.length === 0) {
-            const diffHtml = renderHunksReadonly(lines || []);
-            const label = diffHtml ? `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Changes</div></div></div>` : '';
-            diffEl.innerHTML = `
-    <div class="hunk">
-      <div class="hline"><div class="gutter">commit</div><div class="code">${escapeHtml(commit.id || '')}</div></div>
-      <div class="hline"><div class="gutter">Author</div><div class="code">${escapeHtml(commit.author || 'You <you@example.com>')}</div></div>
-      <div class="hline"><div class="gutter">Message</div><div class="code">${escapeHtml(commit.msg || '')}</div></div>
-    </div>
-    ${label}${diffHtml || ''}`;
-            return;
-        }
-
-        // Build two-column layout: file list and selected file content
-        const sidebar = `<div class="commit-files" style="width: 280px; flex: 0 0 280px; border-right: 1px solid var(--panel-border, #333); overflow:auto;">
-          ${files.map((f, i) => {
-              const cls = i === 0 ? 'row active' : 'row';
-              const status = (f.status || '').toUpperCase();
-              return `<div class="${cls}" data-idx="${i}"><span class="status ${statusClass(status)}">${escapeHtml(status)}</span><div class="file" title="${escapeHtml(f.path)}">${escapeHtml(f.path)}</div></div>`;
-          }).join('')}
-        </div>`;
-        const right = `<div class="commit-right" style="flex:1; overflow:auto; padding-left: 8px; display:flex; flex-direction:column;"><div class="commit-content">${renderHunksReadonly(files[0].lines)}</div></div>`;
-
-        diffEl.innerHTML = `
-    <div class="hunk">
-      <div class="hline"><div class="gutter">commit</div><div class="code">${escapeHtml(commit.id || '')}</div></div>
-      <div class="hline"><div class="gutter">Author</div><div class="code">${escapeHtml(commit.author || 'You <you@example.com>')}</div></div>
-      <div class="hline"><div class="gutter">Message</div><div class="code">${escapeHtml(commit.msg || '')}</div></div>
-    </div>
-    <div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">${files.length} file${files.length===1?'':'s'} changed</div></div></div>
-    <div class="commit-diff" style="display:flex; min-height: 240px; gap: 8px;">${sidebar}${right}</div>`;
-
-        // Sidebar interactions
-        const sideEl = diffEl.querySelector('.commit-files');
-        const contentEl = diffEl.querySelector('.commit-content');
-        if (sideEl && contentEl) {
-            sideEl.querySelectorAll<HTMLElement>('.row').forEach(row => {
-                row.addEventListener('click', () => {
-                    sideEl.querySelectorAll('.row').forEach(r => r.classList.remove('active'));
-                    row.classList.add('active');
-                    const idx = Number(row.getAttribute('data-idx') || '-1');
-                    if (idx >= 0 && idx < files.length) {
-                        (contentEl as HTMLElement).innerHTML = renderHunksReadonly(files[idx].lines);
-                    }
-                });
-            });
-        }
-    } catch (e) {
-        console.warn('git_diff_commit failed', e);
-        // Keep header; show error line
-        diffEl.innerHTML += `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Failed to load diff</div></div></div>`;
-    }
-}
-
-/* ---------------- hydration ---------------- */
-
-/** Keep names consistent with backend: git_list_branches / git_current_branch */
-export async function hydrateBranches() {
-    if (!TAURI.has) return;
-    try {
-        const list = await TAURI.invoke<any[]>('git_list_branches');
-        const head = await TAURI.invoke<{ detached: boolean; branch?: string; commit?: string }>('git_head_status').catch(() => ({ detached: false } as any));
-
-        const has = Array.isArray(list) && list.length > 0;
-        state.hasRepo = state.hasRepo || has; // don’t flip to false if another hydrate confirms true
-
-        if (has) {
-            state.branches = list as any;
-            state.branch = (head as any)?.branch || (list.find((b: any) => b.current)?.name) || state.branch || 'main';
-            window.dispatchEvent(new CustomEvent('app:branches-updated'));
-        }
-    } catch (e) {
-        // Don’t nuke state here; status/summary calls will decide hasRepo
-        console.warn('hydrateBranches failed', e);
-    }
-}
-
-/** Status drives file list; on failure we clear files but don’t assert repo absence unless it’s consistent */
-export async function hydrateStatus() {
-    if (!TAURI.has) return;
-    try {
-        const result = await TAURI.invoke<{ files: any[]; ahead?: number; behind?: number }>('git_status');
-        state.hasRepo = true;
-        state.files = Array.isArray(result?.files) ? (result.files as any) : [];
-        // Default-select all files unless the user has modified selection
-        const currentPaths = new Set((state.files || []).map(f => f.path));
-        if (state.defaultSelectAll) {
-            state.selectedFiles = new Set(Array.from(currentPaths));
-        } else {
-            // prune stale selections no longer present
-            state.selectedFiles.forEach(p => { if (!currentPaths.has(p)) state.selectedFiles.delete(p); });
-        }
-        // ahead/behind are optional in older backends; default to 0
-        (state as any).ahead = Number((result as any)?.ahead || 0);
-        (state as any).behind = Number((result as any)?.behind || 0);
-        renderList();
-        window.dispatchEvent(new CustomEvent('app:status-updated'));
-    } catch (e) {
-        console.warn('hydrateStatus failed', e);
-        state.files = [];
-        state.selectedFiles.clear();
-        renderList();
-        window.dispatchEvent(new CustomEvent('app:status-updated'));
-    }
-}
-
-export async function hydrateCommits() {
-    if (!TAURI.has) return;
-    try {
-        const list = await TAURI.invoke<any[]>('git_log', { limit: 100 });
-        state.hasRepo = true;
-        state.commits = Array.isArray(list) ? (list as any) : [];
-        const aheadCount = Number((state as any).ahead || 0);
-        if (aheadCount > 0) {
-            try {
-                const aheadList = await TAURI.invoke<any[]>('git_log', { limit: 1000, rev: '@{upstream}..HEAD' });
-                const ids = new Set<string>();
-                (aheadList || []).forEach((c: any) => { if (c?.id) ids.add(String(c.id)); });
-                (state as any).aheadIds = ids;
-            } catch {
-                (state as any).aheadIds = new Set<string>();
-            }
-        } else {
-            (state as any).aheadIds = new Set<string>();
-        }
-        if (prefs.tab === 'history') renderList();
-    } catch (e) {
-        console.warn('hydrateCommits failed', e);
-        state.commits = [];
-    }
-}
-
-export async function hydrateStash() {
-    if (!TAURI.has) return;
-    try {
-        const list = await TAURI.invoke<any[]>('git_stash_list');
-        (state as any).stash = Array.isArray(list) ? (list as any) : [];
-        if (prefs.tab === 'stash') renderList();
-    } catch (e) {
-        console.warn('hydrateStash failed', e);
-        (state as any).stash = [];
-    }
-}
-
-function renderHunksWithSelection(lines: string[]) {
-    if (!lines || !lines.length) return '';
-    // Find hunks; hide prelude lines like 'diff --git' / 'index ...' / '---' / '+++'
-    let idx = lines.findIndex(l => l.startsWith('@@'));
-    const rest = idx >= 0 ? lines.slice(idx) : [];
-    const starts: number[] = [];
-    rest.forEach((l, i) => { if (l.startsWith('@@')) starts.push(i); });
-    starts.push(rest.length);
-
-    if (starts.length <= 1) {
-        return `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">No textual hunks to display</div></div></div>`;
-    }
-
-    let html = '';
-    // Render each hunk with a compact checkbox in the gutter
-    for (let h = 0; h < starts.length - 1; h++) {
-        const s = starts[h];
-        const e = starts[h+1];
-        const hunkLines = rest.slice(s, e);
-        const offset = (idx >= 0 ? idx : 0) + s; // approximate numbering
-        const body = hunkLines.map((ln, i) => {
-            const first = (typeof ln === 'string' ? ln[0] : ' ') || ' ';
-            const isChange = first === '+' || first === '-';
-            const lineCheckbox = isChange ? `<label class="pick-toggle"><input type="checkbox" class="pick-line" data-hunk="${h}" data-line="${i}" /><span class="sr-only">Include line</span></label>` : '';
-            const t = first === '+' ? 'add' : first === '-' ? 'del' : '';
-            return `<div class="hline ${t}"><div class="gutter">${lineCheckbox}${offset + i + 1}</div><div class="code">${escapeHtml(String(ln))}</div></div>`;
-        }).join('');
-        html += `<div class="hunk" data-hunk-index="${h}">
-  <div class="hline"><div class="gutter"><label class="pick-toggle"><input type="checkbox" class="pick-hunk" data-hunk="${h}" /><span class="sr-only">Include hunk</span></label></div><div class="code"></div></div>
-  ${body}
-</div>`;
-    }
-    return html;
-}
-
-function renderHunksReadonly(lines: string[]) {
-    if (!lines || !lines.length) return '';
-    let idx = lines.findIndex(l => l.startsWith('@@'));
-    const rest = idx >= 0 ? lines.slice(idx) : [];
-    const starts: number[] = [];
-    rest.forEach((l, i) => { if (l.startsWith('@@')) starts.push(i); });
-    starts.push(rest.length);
-    if (starts.length <= 1) {
-        return `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">No textual hunks to display</div></div></div>`;
-    }
-
-    let html = '';
-    for (let h = 0; h < starts.length - 1; h++) {
-        const s = starts[h];
-        const e = starts[h+1];
-        const hunkLines = rest.slice(s, e);
-        const offset = (idx >= 0 ? idx : 0) + s;
-        html += `<div class="hunk">${hunkLines.map((ln, i) => hline(ln, offset + i + 1)).join('')}</div>`;
-    }
-    return html;
-}
-
-function hline(ln: string, n: number) {
-    const first = (typeof ln === 'string' ? ln[0] : ' ') || ' ';
-    const t = first === '+' ? 'add' : first === '-' ? 'del' : '';
-    return `<div class="hline ${t}"><div class="gutter">${n}</div><div class="code">${escapeHtml(String(ln))}</div></div>`;
-}
-
-function showStashFooter() {
-    if (!leftFootEl) return;
-    const foot = ensureStashFooterControls();
-    if (!foot) return;
-    leftFootEl.dataset.mode = 'stash';
-    leftFootEl.classList.add('show');
-    if (undoLeftBtn) undoLeftBtn.style.display = 'none';
-    foot.classList.add('show');
-}
-
-function hideStashFooter() {
-    if (!leftFootEl) return;
-    if (leftFootEl.dataset.mode === 'stash') {
-        leftFootEl.classList.remove('show');
-        leftFootEl.dataset.mode = '';
-    }
-    if (undoLeftBtn) undoLeftBtn.style.display = '';
-    if (stashFootEl) stashFootEl.classList.remove('show');
-}
-
-function ensureStashFooterControls(): HTMLElement | null {
-    if (!leftFootEl) return null;
-    if (!stashFootEl) {
-        stashFootEl = document.createElement('div');
-        stashFootEl.id = 'stash-foot-controls';
-        stashFootEl.className = 'stash-foot';
-        stashFootEl.innerHTML = `
-          <button class="btn" id="stash-create-btn" title="Stash current changes">Create Stash</button>
-          <button class="btn" id="stash-apply-btn" disabled>Apply</button>
-          <button class="btn" id="stash-pop-btn" disabled>Pop</button>
-          <button class="btn" id="stash-drop-btn" disabled>Drop</button>
-        `;
-        leftFootEl.appendChild(stashFootEl);
-    }
-    if (!stashFootBound && stashFootEl) {
-        wireStashFooterButtons(stashFootEl);
-        stashFootBound = true;
-    }
-    return stashFootEl;
-}
-
-function wireStashFooterButtons(container: HTMLElement) {
-    const createBtn = container.querySelector<HTMLButtonElement>('#stash-create-btn');
-    createBtn?.addEventListener('click', () => {
-        openStashConfirm({
-            onSuccess: async () => {
-                await Promise.allSettled([hydrateStatus(), hydrateStash()]);
-                renderList();
-            },
-        });
-    });
-
-    const applyBtn = container.querySelector<HTMLButtonElement>('#stash-apply-btn');
-    applyBtn?.addEventListener('click', async () => {
-        const selector = getActiveStashSelector();
-        if (!selector) return;
-        try {
-            if (!TAURI.has) return;
-            await TAURI.invoke('git_stash_apply', { selector });
-            notify('Applied stash');
-            await Promise.allSettled([hydrateStatus(), hydrateStash()]);
-            renderList();
-        } catch (e) { console.warn('git_stash_apply failed', e); notify('Failed to apply stash'); }
-    });
-
-    const popBtn = container.querySelector<HTMLButtonElement>('#stash-pop-btn');
-    popBtn?.addEventListener('click', async () => {
-        const selector = getActiveStashSelector();
-        if (!selector) return;
-        try {
-            if (!TAURI.has) return;
-            await TAURI.invoke('git_stash_pop', { selector });
-            notify('Popped stash');
-            await Promise.allSettled([hydrateStatus(), hydrateStash()]);
-            renderList();
-        } catch (e) { console.warn('git_stash_pop failed', e); notify('Failed to pop stash'); }
-    });
-
-    const dropBtn = container.querySelector<HTMLButtonElement>('#stash-drop-btn');
-    dropBtn?.addEventListener('click', async () => {
-        const selector = getActiveStashSelector();
-        if (!selector) return;
-        const ok = window.confirm(`Drop ${selector}? This cannot be undone.`);
-        if (!ok) return;
-        try {
-            if (!TAURI.has) return;
-            await TAURI.invoke('git_stash_drop', { selector });
-            notify('Dropped stash');
-            state.currentStash = '';
-            await Promise.allSettled([hydrateStash()]);
-            renderList();
-        } catch (e) { console.warn('git_stash_drop failed', e); notify('Failed to drop stash'); }
-    });
-}
-
-// Group commit diff into per-file blocks based on `diff --git` markers.
-function parseCommitDiffByFile(lines: string[]): { path: string; status: string; lines: string[] }[] {
-    if (!Array.isArray(lines) || lines.length === 0) return [];
-    const files: { path: string; status: string; lines: string[] }[] = [];
-    let i = 0;
-    while (i < lines.length) {
-        const l = lines[i] || '';
-        if (l.startsWith('diff --git ')) {
-            const parts = l.split(' ');
-            const aPath = parts[2] || '';
-            const bPath = parts[3] || '';
-            const pathA = aPath.startsWith('a/') ? aPath.slice(2) : aPath;
-            const pathB = bPath.startsWith('b/') ? bPath.slice(2) : bPath;
-            const path = pathB || pathA;
-
-            // Collect until next file header or end
-            let j = i + 1;
-            while (j < lines.length && !String(lines[j]).startsWith('diff --git ')) j++;
-            const block = lines.slice(i, j);
-
-            const prelude = block.slice(0, Math.min(block.length, 10));
-            const isAdd = prelude.some(s => s.startsWith('new file mode') || s.startsWith('--- /dev/null'));
-            const isDel = prelude.some(s => s.startsWith('deleted file mode') || s.startsWith('+++ /dev/null'));
-            const status = isAdd ? 'A' : isDel ? 'D' : 'M';
-
-            files.push({ path, status, lines: block });
-            i = j;
-            continue;
-        }
-        i++;
-    }
-    return files;
-}
-
-function onFileClick(e: MouseEvent, file: { path: string }, index: number, visible: { path: string }[]) {
-    if (suppressNextClick) {
-        suppressNextClick = false;
-        if (state.diffSelectedFiles && state.diffSelectedFiles.size > 1) {
-            highlightRow(index);
-            renderCombinedDiff(Array.from(state.diffSelectedFiles));
-        } else {
-            clearDiffSelection();
-            selectFile(file, index);
-        }
-        return;
-    }
-    const isToggle = e.ctrlKey || e.metaKey;
-    const isRange = e.shiftKey && lastClickedIndex >= 0;
-
-    if (isRange) {
-        // Toggle (invert) selection for the whole range
-        const a = Math.min(lastClickedIndex, index);
-        const b = Math.max(lastClickedIndex, index);
-        for (let i = a; i <= b; i++) {
-            const p = visible[i]?.path; if (!p) continue;
-            if (state.selectedFiles.has(p)) state.selectedFiles.delete(p);
-            else state.selectedFiles.add(p);
-        }
-        state.defaultSelectAll = false;
-        updateSelectAllState(visible);
-        renderList();
-        // Keep the clicked file as the viewed file
-        selectFile(file, index);
-    } else if (isToggle) {
-        const on = !state.selectedFiles.has(file.path);
-        toggleFilePick(file.path, on);
-        updateSelectAllState(visible);
-        // Update row UI immediately (checkbox + picked class)
-        if (listEl) {
-            const sel = `li.row[data-path="${(file.path || '').replace(/([\"\\])/g, '\\$1')}"]`;
-            const row = listEl.querySelector<HTMLElement>(sel);
-            if (row) {
-                row.classList.toggle('picked', on);
-                const cb = row.querySelector<HTMLInputElement>('input.pick');
-                if (cb) { cb.checked = on; (cb as any).indeterminate = false; }
-            }
-        }
-        // Do not affect diff multi-selection with commit toggle
-        if (!(state.diffSelectedFiles && state.diffSelectedFiles.size > 1)) {
-            selectFile(file, index);
-        }
-    } else {
-        // plain click: clear diff multi-selection and view just this file
-        clearDiffSelection();
-        selectFile(file, index);
-    }
-    lastClickedIndex = index;
-    updateCommitButton();
-}
-
-function onFileMouseDown(e: MouseEvent, file: { path: string }, index: number, visible: { path: string }[], li: HTMLElement) {
-    if (e.button !== 0) return; // left only
-    const mode = e.shiftKey ? 'diff' : (e.ctrlKey || e.metaKey) ? 'commit' : null;
-    if (mode === null) {
-        // Plain click → let the click handler run without drag side-effects
-        dragMode = null;
-        isDragSelecting = false;
-        dragMoved = false;
-        dragVisited.clear();
-        return;
-    }
-
-    e.preventDefault(); // avoid starting native selection
-    dragMode = mode;
-    dragMoved = false;
-    isDragSelecting = true;
-    dragVisited.clear();
-    document.body.classList.add('drag-selecting');
-    // Clear any current selection
-    try { const sel = window.getSelection?.(); sel && sel.removeAllRanges(); } catch {}
-    if (dragMode === 'diff') {
-        clearActiveRows();
-        dragTargetState = true;
-        dragStartIndex = index; dragCurrentIndex = index;
-        dragPreDiff = new Set(state.diffSelectedFiles);
-        updateDragRange(visible);
-    } else if (dragMode === 'commit') {
-        const currentlyOn = state.selectedFiles.has(file.path);
-        dragTargetState = !currentlyOn;
-        dragStartIndex = index; dragCurrentIndex = index;
-        dragPrePicked = new Set(state.selectedFiles);
-        updateDragRange(visible);
-    }
-
-    const startX = e.clientX, startY = e.clientY;
-    const onMove = (mv: MouseEvent) => {
-        if (!dragMoved && (Math.abs(mv.clientX - startX) + Math.abs(mv.clientY - startY) > 3)) dragMoved = true;
-        const el = document.elementFromPoint(mv.clientX, mv.clientY) as HTMLElement | null;
-        const row = el ? el.closest('li.row[data-path]') as HTMLElement | null : null;
-        if (row) {
-            const p = row.getAttribute('data-path') || '';
-            const i2 = visible.findIndex(v => v.path === p);
-            if (i2 >= 0 && i2 !== dragCurrentIndex) {
-                dragCurrentIndex = i2;
-                updateDragRange(visible);
-            }
-        }
-    };
-    const onUp = () => {
-        document.removeEventListener('mouseup', onUp);
-        document.removeEventListener('mousemove', onMove);
-        isDragSelecting = false;
-        document.body.classList.remove('drag-selecting');
-        updateSelectAllState(visible);
-        updateCommitButton();
-        if (dragMoved) suppressNextClick = true;
-        lastClickedIndex = index;
-        // If multiple files are diff-selected, render combined view
-        if (state.diffSelectedFiles && state.diffSelectedFiles.size > 1) {
-            renderCombinedDiff(Array.from(state.diffSelectedFiles));
-        }
-        dragMode = null;
-    };
-    document.addEventListener('mousemove', onMove);
-    document.addEventListener('mouseup', onUp, { once: true });
-}
-
-function applySelect(path: string, on: boolean, rowEl: HTMLElement | null, visible: { path: string }[], mode: 'diff'|'commit') {
-    state.defaultSelectAll = false;
-    if (mode === 'commit') {
-        if (on) state.selectedFiles.add(path); else state.selectedFiles.delete(path);
-        if (rowEl) rowEl.classList.toggle('picked', on);
-        const selector = `li.row[data-path="${path.replace(/([\"\\])/g, '\\$1')}"] input.pick`;
-        const cb = listEl?.querySelector<HTMLInputElement>(selector) || null;
-        if (cb) { cb.checked = on; (cb as any).indeterminate = false; }
-    } else {
-        if (on) state.diffSelectedFiles.add(path); else state.diffSelectedFiles.delete(path);
-        if (rowEl) rowEl.classList.toggle('diffsel', on);
-    }
-}
-
-function updateDragRange(visible: { path: string }[]) {
-    if (!isDragSelecting || dragMode === null) return;
-    const a = Math.min(dragStartIndex, dragCurrentIndex);
-    const b = Math.max(dragStartIndex, dragCurrentIndex);
-    if (dragMode === 'diff') {
-        const next = new Set(dragPreDiff);
-        for (let i = 0; i < visible.length; i++) {
-            const p = visible[i]?.path; if (!p) continue;
-            if (i >= a && i <= b) next.add(p); else if (!dragPreDiff.has(p)) next.delete(p);
-        }
-        state.diffSelectedFiles = next;
-        if (listEl) {
-            visible.forEach(v => {
-                const row = listEl!.querySelector<HTMLElement>(`li.row[data-path="${(v.path || '').replace(/([\"\\])/g, '\\$1')}"]`);
-                if (row) row.classList.toggle('diffsel', state.diffSelectedFiles.has(v.path));
-            });
-        }
-    } else if (dragMode === 'commit') {
-        const next = new Set<string>();
-        for (let i = 0; i < visible.length; i++) {
-            const p = visible[i]?.path; if (!p) continue;
-            const inRange = i >= a && i <= b;
-            const on = inRange ? dragTargetState : dragPrePicked.has(p);
-            if (on) next.add(p);
-            if (listEl) {
-                const row = listEl!.querySelector<HTMLElement>(`li.row[data-path="${(p || '').replace(/([\"\\])/g, '\\$1')}"]`);
-                if (row) row.classList.toggle('picked', on);
-                const cb = listEl!.querySelector<HTMLInputElement>(`li.row[data-path="${(p || '').replace(/([\"\\])/g, '\\$1')}"] input.pick`);
-                if (cb) { cb.checked = on; (cb as any).indeterminate = false; }
-            }
-            // If this is the currently viewed file, mirror commit selection to hunk selection
-            if (state.currentFile && p === state.currentFile) {
-                if (on) {
-                    state.selectedHunks = allHunkIndices(state.currentDiff);
-                    (state as any).selectedHunksByFile[state.currentFile] = state.selectedHunks.slice();
-                } else {
-                    state.selectedHunks = [];
-                    delete (state as any).selectedHunksByFile[state.currentFile];
-                }
-                updateHunkCheckboxes();
-            }
-        }
-        state.selectedFiles = next;
-    }
-}
-
-
-function onFileContextMenu(ev: MouseEvent, f: { path: string }) {
-    ev.preventDefault();
-    const x = ev.clientX, y = ev.clientY;
-    const totalFiles = Array.isArray(state.files) ? state.files.length : 0;
-    const selectedPaths = Array.from(state.selectedFiles || []);
-    const hasManualSelection = (!state.defaultSelectAll) || (selectedPaths.length > 0 && selectedPaths.length !== totalFiles);
-    const manualSelection = hasManualSelection ? selectedPaths : [];
-    const clickedInSelection = manualSelection.includes(f.path);
-    const hasMultiSelection = manualSelection.length > 1 && clickedInSelection;
-    const hasSingleSelection = manualSelection.length === 1 && clickedInSelection;
-    const items: CtxItem[] = [];
-    const openStashForPaths = (paths: string[], defaultMessage: string) => {
-        if (!paths.length) return;
-        openStashConfirm({
-            defaultMessage,
-            includeUntracked: false,
-            paths,
-            onSuccess: async () => {
-                await Promise.allSettled([hydrateStatus(), hydrateStash()]);
-                renderList();
-            },
-        });
-    };
-    items.push({ label: 'Discard changes', action: async () => {
-        if (!TAURI.has) return;
-        const ok = window.confirm(`Discard all changes in \n${f.path}? This cannot be undone.`);
-        if (!ok) return;
-        try { await TAURI.invoke('git_discard_paths', { paths: [f.path] }); await Promise.allSettled([hydrateStatus()]); }
-        catch { notify('Discard failed'); }
-    }});
-    if (hasManualSelection && clickedInSelection) {
-        items.push({ label: 'Discard selected files', action: async () => {
-            if (!TAURI.has) return;
-            const paths = manualSelection.slice();
-            const ok = window.confirm(`Discard all changes in ${paths.length} selected file(s)? This cannot be undone.`);
-            if (!ok) return;
-            try { await TAURI.invoke('git_discard_paths', { paths }); await Promise.allSettled([hydrateStatus()]); }
-            catch { notify('Discard failed'); }
-        }});
-        if (hasMultiSelection) {
-            items.push({ label: 'Create stash from selection…', action: () => {
-                openStashForPaths(manualSelection.slice(), 'WIP selection');
-            }});
-        }
-    }
-    const singleTarget = hasSingleSelection ? manualSelection[0] : f.path;
-    const defaultMsg = `WIP ${singleTarget}`;
-    items.push({ label: 'Create stash for this file…', action: () => {
-        openStashForPaths([singleTarget], defaultMsg);
-    }});
-    items.push({ label: '---' });
-    items.push({ label: 'Track with Git LFS', action: () => {
-        if (!TAURI.has) {
-            notify('Git LFS is available in the desktop app');
-            return;
-        }
-        const targets = (hasManualSelection && clickedInSelection ? manualSelection.slice() : [f.path]).filter(Boolean);
-        if (!targets.length) return;
-        (async () => {
-            try {
-                await TAURI.invoke('git_lfs_track_paths', { paths: targets });
-                notify(targets.length > 1 ? 'Tracked files with Git LFS' : 'Tracked file with Git LFS');
-                await Promise.allSettled([hydrateStatus()]);
-            } catch {
-                notify('Git LFS track failed');
-            }
-        })();
-    }});
-    buildCtxMenu(items, x, y);
-}
-
-function toggleFilePick(path: string, on: boolean) {
-    if (!path) return;
-    state.defaultSelectAll = false;
-    if (on) state.selectedFiles.add(path); else state.selectedFiles.delete(path);
-    // If this is the currently viewed file, mirror selection to all hunks in the diff
-    if (state.currentFile && state.currentFile === path) {
-        if (on) {
-            // Select all hunks
-            state.selectedHunks = allHunkIndices(state.currentDiff);
-            (state as any).selectedHunksByFile[state.currentFile] = state.selectedHunks.slice();
-            // Select all changed lines per hunk
-            const root = diffEl as HTMLElement;
-            const rec: Record<number, number[]> = {};
-            const lineBoxes = root.querySelectorAll<HTMLInputElement>('input.pick-line');
-            lineBoxes.forEach(b => {
-                const h = Number(b.dataset.hunk || -1);
-                const l = Number(b.dataset.line || -1);
-                if (h < 0 || l < 0) return;
-                (rec[h] ||= []).push(l);
-                b.checked = true;
-            });
-            // Normalize and assign
-            Object.keys(rec).forEach(k => { rec[Number(k)] = Array.from(new Set(rec[Number(k)])).sort((a,b)=>a-b); });
-            (state as any).selectedLinesByFile[state.currentFile] = rec;
-        } else {
-            state.selectedHunks = [];
-            delete (state as any).selectedHunksByFile[state.currentFile];
-            // Clear all line selections
-            const root = diffEl as HTMLElement;
-            const lineBoxes = root.querySelectorAll<HTMLInputElement>('input.pick-line');
-            lineBoxes.forEach(b => { b.checked = false; });
-            delete (state as any).selectedLinesByFile[state.currentFile];
-        }
-        updateHunkCheckboxes();
-    }
-    updateCommitButton();
-}
-
-function updateSelectAllState(visible: { path: string }[]) {
-    if (!selectAllBox) return;
-    if (prefs.tab !== 'changes') {
-        selectAllBox.indeterminate = false;
-        selectAllBox.checked = false;
-        return;
-    }
-    const total = visible.length;
-    if (total === 0) {
-        selectAllBox.indeterminate = false;
-        selectAllBox.checked = false;
-        return;
-    }
-    const selected = visible.filter(f => state.selectedFiles.has(f.path)).length;
-    selectAllBox.indeterminate = selected > 0 && selected < total;
-    selectAllBox.checked = selected === total;
-}
-
-function getVisibleFiles(): { path: string }[] {
-    if (prefs.tab !== 'changes') return [];
-    const q = (filterInput?.value || '').trim().toLowerCase();
-    return (state.files || []).filter(f => !q || (f.path || '').toLowerCase().includes(q));
-}
-
-function toggleSelectAll(on: boolean, visible: { path: string }[]) {
-    if (on) {
-        visible.forEach(f => { if (f.path) toggleFilePick(f.path, true); });
-    } else {
-        visible.forEach(f => { if (f.path) toggleFilePick(f.path, false); });
-    }
-}
-
-function bindHunkToggles(root: HTMLElement) {
-    const boxes = root.querySelectorAll<HTMLInputElement>('input.pick-hunk');
-    boxes.forEach(b => {
-        b.addEventListener('change', () => {
-            state.defaultSelectAll = false;
-            const idx = Number(b.dataset.hunk || -1);
-            if (b.checked) {
-                if (!state.selectedHunks.includes(idx)) state.selectedHunks.push(idx);
-            } else {
-                state.selectedHunks = state.selectedHunks.filter(i => i !== idx);
-            }
-            if (state.currentFile) {
-                (state as any).selectedHunksByFile[state.currentFile] = state.selectedHunks.slice();
-                // Toggle all line checkboxes in this hunk accordingly
-                const linesInHunk = Array.from(root.querySelectorAll<HTMLInputElement>(`input.pick-line[data-hunk="${idx}"]`));
-                const rec: Record<number, number[]> = (state as any).selectedLinesByFile[state.currentFile] || {};
-                if (b.checked) {
-                    const picked: number[] = [];
-                    linesInHunk.forEach((el) => { el.checked = true; picked.push(Number(el.dataset.line || -1)); });
-                    rec[idx] = Array.from(new Set([...(rec[idx] || []), ...picked])).sort((a,b)=>a-b);
-                } else {
-                    linesInHunk.forEach((el) => { el.checked = false; });
-                    delete rec[idx];
-                }
-                (state as any).selectedLinesByFile[state.currentFile] = rec;
-            }
-            // Update file checkbox and selectedFiles based on hunk selection
-            const before = (state.selectedHunks || []).length;
-            // Sync checkbox tri-state
-            (function(){ syncFileCheckboxWithHunks(); })();
-            if ((state.selectedHunks || []).length === 0 && state.currentFile) {
-                state.selectedFiles.delete(state.currentFile);
-                delete (state as any).selectedHunksByFile[state.currentFile];
-            }
-            updateSelectAllState(getVisibleFiles());
-            updateCommitButton();
-            const hk = b.closest('.hunk') as HTMLElement | null;
-            if (hk) hk.classList.toggle('picked', b.checked);
-        });
-    });
-
-    // Per-line toggles
-    const lineBoxes = root.querySelectorAll<HTMLInputElement>('input.pick-line');
-    lineBoxes.forEach(b => {
-        b.addEventListener('change', () => {
-            state.defaultSelectAll = false;
-            const hunk = Number(b.dataset.hunk || -1);
-            const line = Number(b.dataset.line || -1);
-            if (!state.currentFile || hunk < 0 || line < 0) return;
-            const rec: Record<number, number[]> = (state as any).selectedLinesByFile[state.currentFile] || {};
-            const cur = new Set<number>(rec[hunk] || []);
-            if (b.checked) cur.add(line); else cur.delete(line);
-            rec[hunk] = Array.from(cur).sort((a,b)=>a-b);
-            if (rec[hunk].length === 0) delete rec[hunk];
-            (state as any).selectedLinesByFile[state.currentFile] = rec;
-
-            // update hunk checkbox state
-            const hunkBox = root.querySelector<HTMLInputElement>(`input.pick-hunk[data-hunk="${hunk}"]`);
-            if (hunkBox) {
-                const total = root.querySelectorAll<HTMLInputElement>(`input.pick-line[data-hunk="${hunk}"]`).length;
-                const sel = rec[hunk]?.length || 0;
-                (hunkBox as any).indeterminate = sel > 0 && sel < total;
-                hunkBox.checked = sel === total && total > 0;
-                // Reflect fully-selected lines into selectedHunks for consistency
-                const idx = Number(hunkBox.dataset.hunk || -1);
-                if (sel === total && total > 0) {
-                    if (!state.selectedHunks.includes(idx)) state.selectedHunks.push(idx);
-                } else {
-                    state.selectedHunks = state.selectedHunks.filter(i => i !== idx);
-                }
-                // Persist hunk selection snapshot
-                if (state.currentFile) {
-                    (state as any).selectedHunksByFile[state.currentFile] = state.selectedHunks.slice();
-                }
-            }
-            // Update file row checkbox tri-state
-            syncFileCheckboxWithHunks();
-            updateSelectAllState(getVisibleFiles());
-            updateCommitButton();
-        });
-    });
-}
-
-function syncFileCheckboxWithHunks() {
-    if (!state.currentFile) return;
-    const totalHunks = allHunkIndices(state.currentDiff).length;
-    const selHunks = (state.selectedHunks || []).length;
-
-    if (totalHunks === 0) {
-        updateListCheckboxForPath(state.currentFile, false, false);
-        state.selectedFiles.delete(state.currentFile);
-        return;
-    }
-
-    // Consider per-line partial selections for this file
-    const rec: Record<number, number[]> = (state as any).selectedLinesByFile[state.currentFile] || {};
-    // Build per-hunk change-line counts from current diff
-    const lines = state.currentDiff || [];
-    const first = lines.findIndex(l => (l || '').startsWith('@@'));
-    const rest = first >= 0 ? lines.slice(first) : [];
-    const starts: number[] = [];
-    for (let i = 0; i < rest.length; i++) { if ((rest[i] || '').startsWith('@@')) starts.push(i); }
-    starts.push(rest.length);
-    const changeCounts: number[] = [];
-    for (let h = 0; h < Math.max(0, starts.length - 1); h++) {
-        const s = starts[h];
-        const e = starts[h + 1];
-        const block = rest.slice(s + 1, e); // skip header
-        const cnt = block.reduce((acc, ln) => {
-            const ch = (ln || '')[0] || ' ';
-            return acc + ((ch === '+' || ch === '-') ? 1 : 0);
-        }, 0);
-        changeCounts[h] = cnt;
-    }
-
-    const hasAnyLineSel = Object.keys(rec).length > 0;
-    const hasPartialLineSel = Object.keys(rec).some(k => {
-        const h = Number(k);
-        const chosen = Array.isArray(rec[h]) ? rec[h].length : 0;
-        const total = changeCounts[h] || 0;
-        return chosen > 0 && chosen < total;
-    });
-
-    // Determine file checkbox state
-    if (selHunks === totalHunks && !hasPartialLineSel) {
-        updateListCheckboxForPath(state.currentFile, true, false);
-        state.selectedFiles.add(state.currentFile);
-    } else if (selHunks === 0 && !hasAnyLineSel) {
-        updateListCheckboxForPath(state.currentFile, false, false);
-        state.selectedFiles.delete(state.currentFile);
-    } else {
-        updateListCheckboxForPath(state.currentFile, false, true);
-        state.selectedFiles.delete(state.currentFile);
-    }
-}
-
-function allHunkIndices(lines: string[]) {
-    if (!Array.isArray(lines) || !lines.length) return [] as number[];
-    const idx = lines.findIndex(l => l.startsWith('@@'));
-    const rest = idx >= 0 ? lines.slice(idx) : [];
-    const starts: number[] = [];
-    rest.forEach((l, i) => { if (l.startsWith('@@')) starts.push(i); });
-    // indices are 0..(count-1)
-    return starts.map((_, i) => i);
-}
-
-function updateHunkCheckboxes() {
-    const root = diffEl as HTMLElement;
-    if (!root) return;
-    const boxes = root.querySelectorAll<HTMLInputElement>('input.pick-hunk');
-    boxes.forEach(b => {
-        const idx = Number(b.dataset.hunk || -1);
-        const on = state.selectedHunks.includes(idx);
-        b.checked = on;
-        const hk = b.closest('.hunk') as HTMLElement | null;
-        if (hk) hk.classList.toggle('picked', on);
-    });
-
-    // Sync line checkboxes from state
-    if (state.currentFile) {
-        const rec: Record<number, number[]> = (state as any).selectedLinesByFile[state.currentFile] || {};
-        const lboxes = root.querySelectorAll<HTMLInputElement>('input.pick-line');
-        lboxes.forEach(b => {
-            const h = Number(b.dataset.hunk || -1);
-            const l = Number(b.dataset.line || -1);
-            const on = Array.isArray(rec[h]) && rec[h].includes(l);
-            b.checked = on;
-        });
-    }
-}
-
-function updateListCheckboxForPath(path: string, checked: boolean, indeterminate: boolean) {
-    if (!listEl || !path) return;
-    const selector = `li.row[data-path="${path.replace(/(["\\])/g, '\\$1')}"] input.pick`;
-    const cb = listEl.querySelector<HTMLInputElement>(selector);
-    if (cb) {
-        cb.checked = checked;
-        (cb as any).indeterminate = indeterminate;
-    }
-}
-
-async function renderCombinedDiff(paths: string[]) {
-    if (!diffHeadPath || !diffEl) return;
-    clearActiveRows();
-    const files = Array.from(new Set(paths)).filter(Boolean);
-    diffHeadPath.textContent = `Multiple files (${files.length})`;
-    diffEl.innerHTML = '<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">Loading…</div></div></div>';
-    let html = '';
-    for (const p of files) {
-        try {
-            const lines = TAURI.has ? await TAURI.invoke<string[]>('git_diff_file', { path: p }) : [];
-            html += `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">${escapeHtml(p)}</div></div></div>`;
-            html += renderHunksWithSelection(lines || []);
-        } catch {
-            html += `<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">${escapeHtml(p)} (failed to load diff)</div></div></div>`;
-        }
-    }
-    diffEl.innerHTML = html || '<div class="hunk"><div class="hline"><div class="gutter"></div><div class="code">No diffs</div></div></div>';
-}
-
-function clearDiffSelection() {
-    if (!listEl) return;
-    if (state.diffSelectedFiles && state.diffSelectedFiles.size > 0) {
-        state.diffSelectedFiles.clear();
-        // Remove visual class from rows
-        const rows = listEl.querySelectorAll<HTMLElement>('li.row.diffsel');
-        rows.forEach(r => r.classList.remove('diffsel'));
-    }
-}
-
-function clearActiveRows() {
-    if (!listEl) return;
-    const rows = listEl.querySelectorAll<HTMLElement>('li.row.active');
-    rows.forEach(r => r.classList.remove('active'));
-}
-
-function updateCommitButton() {
-    const btn = document.getElementById('commit-btn') as HTMLButtonElement | null;
-    if (!btn) return;
-    const summary = document.getElementById('commit-summary') as HTMLInputElement | null;
-    const summaryFilled = (summary?.value.trim().length ?? 0) > 0;
-    const hunksSelected = Object.keys((state as any).selectedHunksByFile || {})
-        .some((k) => Array.isArray((state as any).selectedHunksByFile[k]) && (state as any).selectedHunksByFile[k].length > 0);
-    const linesSelected = Object.keys((state as any).selectedLinesByFile || {})
-        .some((k) => !!(state as any).selectedLinesByFile[k] && Object.keys((state as any).selectedLinesByFile[k] || {}).length > 0);
-    const filesSelected = !!(state.selectedFiles && state.selectedFiles.size > 0);
-    btn.disabled = !(summaryFilled && (hunksSelected || linesSelected || filesSelected));
-}
-
-function getActiveStashSelector(): string {
-    if (state.currentStash) return state.currentStash;
-    const active = listEl?.querySelector<HTMLElement>('li.row.commit.active');
-    const sel = active?.dataset.selector || '';
-    if (sel) state.currentStash = sel;
-    return sel;
-}
-
-// Convert an ISO/RFC3339 datetime string into a short relative phrase.
-function formatTimeAgo(isoMaybe: string): string {
-    try {
-        const d = new Date(String(isoMaybe || '').trim());
-        const t = d.getTime();
-        if (!isFinite(t)) return (isoMaybe || '').trim();
-        const now = Date.now();
-        let sec = Math.max(0, Math.round((now - t) / 1000));
-        if (sec < 45) return 'just now';
-        if (sec < 90) return '1 minute ago';
-        let min = Math.round(sec / 60);
-        if (min < 60) return `${min} minute${min === 1 ? '' : 's'} ago`;
-        let hr = Math.round(min / 60);
-        if (hr < 24) return `${hr} hour${hr === 1 ? '' : 's'} ago`;
-        let day = Math.round(hr / 24);
-        if (day === 1) return 'yesterday';
-        if (day < 7) return `${day} day${day === 1 ? '' : 's'} ago`;
-        let wk = Math.round(day / 7);
-        if (wk === 1) return '1 week ago';
-        if (wk < 5) return `${wk} weeks ago`;
-        let mon = Math.round(day / 30);
-        if (mon === 1) return '1 month ago';
-        if (mon < 12) return `${mon} months ago`;
-        let yr = Math.round(day / 365);
-        return `${yr} year${yr === 1 ? '' : 's'} ago`;
-    } catch {
-        return (isoMaybe || '').trim();
-    }
-}
