+++ Frontend/src/scripts/features/repo/filter.ts
@@ -1,0 +1,16 @@
+import { state, prefs } from '../../state/state';
+import { filterInput, selectAllBox } from './context';
+import { renderList } from './list';
+import { getVisibleFiles } from './selectionState';
+import { toggleSelectAll } from './interactions';
+
+export function bindFilter() {
+    filterInput?.addEventListener('input', () => renderList());
+    selectAllBox?.addEventListener('change', () => {
+        if (prefs.tab !== 'changes') return;
+        state.defaultSelectAll = false;
+        const files = getVisibleFiles();
+        toggleSelectAll(Boolean(selectAllBox?.checked), files);
+        renderList();
+    });
+}
